[level 1] Title: 부족한 금액 계산하기, Time: 0.09 ms, Memory: 73.3 MB -BaekjoonHub

+++ 프로그래머스/lv1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -1,12 +1,12 @@
 class Solution {
-    public long solution(int price, long money, int count) {
-        long answer = -1;
+public long solution(int price, int money, int count) {
+long answer = -1;
 
-        for(int i=1; i<=count; i++){
-        money -= price * i;
-        }
-        answer = (money < 0)? (long)money * (-1):0;
+for(int i=1; i<=count; i++){
+money -= price * i;
+}
+answer = (money < 0)? (long)money * (-1):0;
 
-        return answer;
-    }
+return answer;
+}
 }
